backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.0.5
cryptography==41.0.3
defusedxml==0.7.1
Django==2.2.19
django-debug-toolbar==3.2.4
django-extra-fields==0.1
django-filter==2.3.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djoser==2.1.0
drf-base64==2.0
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==5.0.3
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
idna==3.4
isort==5.11.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
oauthlib==3.2.2
pep8-naming==0.13.2
Pillow==9.4.0
postgres==4.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sorl-thumbnail==12.9.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
